# Scorers allocs