--- a/search/scorers/scorer_conjunction_test.go
+++ b/search/scorers/scorer_conjunction_test.go
@@ -1,0 +1,26 @@
+//  Copyright (c) 2016 Couchbase, Inc.
+//  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+//  except in compliance with the License. You may obtain a copy of the License at
+//    http://www.apache.org/licenses/LICENSE-2.0
+//  Unless required by applicable law or agreed to in writing, software distributed under the
+//  License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+//  either express or implied. See the License for the specific language governing permissions
+//  and limitations under the License.
+
+package scorers
+
+import (
+	"testing"
+
+	"github.com/blevesearch/bleve/search"
+)
+
+func BenchmarkConjunctionScore(b *testing.B) {
+	scorer := NewConjunctionQueryScorer(false)
+	constituents := make([]*search.DocumentMatch, 0)
+	constituents = append(constituents, &search.DocumentMatch{ID: "id"})
+	constituents = append(constituents, &search.DocumentMatch{ID: "id"})
+	for i := 0; i < b.N; i++ {
+		scorer.Score(constituents)
+	}
+}