benchmark                       old ns/op     new ns/op     delta
BenchmarkConjunctionScore-2     142           26.4          -81.41%
BenchmarkTermScore-2            1373          1226          -10.71%

benchmark                       old allocs     new allocs     delta
BenchmarkConjunctionScore-2     1              0              -100.00%
BenchmarkTermScore-2            7              6              -14.29%

benchmark                       old bytes     new bytes     delta
BenchmarkConjunctionScore-2     80            0             -100.00%
BenchmarkTermScore-2            792           712           -10.10%