@mschoch 2016-01-15
I have some concerns with this, as there is now state in the scorer which was not the intended design.  I'm not saying it can't work, but that wasn't how it was expected to work.

Also, the integration tests fail.  I'm not sure if its related to that or not.