@kevgs 2016-01-15
I see no problems with properly encapsulated state. But this patch add a restriction on scorer users - it's now disallowed to change value returned by Score(). It's possible to document this restriction on all scorers. The benefit is a better performance. Do you thinks it's worth it?

Integration test failure is probably not related to patch.