@mschoch 2016-01-15
Well the tests are run on every push and have been passing.  However, when run against this pull-request they fail.  So this patch is the most likely cause at the moment.

The part I don't understand is that the same TermQueryScorer is used to score multiple matches.  But, since you only ever allocate one DocumentMatch, aren't you constantly overwriting the same DocumentMatch, and returning references to the same one each time?