@mschoch 2016-02-05
I've reviewed this a final time.  Reducing allocations is great, but unfortunately this change breaks the intended usage of Scorers.  Currently searchers instantiate a single scorer, and use it to score all the hits found by the searcher.  This change only ever creates one DocumentMatch object, and all calls to Score keep modifying this single instance and returning references to it.

The unit tests still pass because they check the contents of the DocumentMatch after each call to Score, but the problem is that we collect all the DocumentMatches and percolate them back up the searcher hierarchy.  This problem is exposed by the integration tests, which is why they are failing.